+++ Makefile
@@ -1,18 +1,18 @@
 .PHONY:	all
 
-backendImage 	= "ube"
-backendName  	= "cube"
-networkName  	= "vault-unsealing"
-moduleImage	= "vufe"
-moduleName	= "vault-unsealing"
+backendImage 	= "consulmod"
+backendName  	= "consulmod"
+networkName  	= "vault-autounsealing"
+moduleImage		= "vaultmod"
+moduleName		= "vaultmod"
 
 infraBuild:
 	docker network create ${networkName}
 
 platformBuild: infraBuild
-	docker build -t ${backendImage} -f tests/platform/backend/Dockerfile.offline tests/platform/backend/
+	docker build -t ${backendImage} -f tests/platform/consul/Dockerfile.offline tests/platform/consul/
 	docker run --network ${networkName} -td -p 8500:8500 --name ${backendName} ${backendImage}
-	docker build -t ${moduleImage} -f tests/platform/frontend/Dockerfile.offline tests/platform/frontend/
+	docker build -t ${moduleImage} -f tests/platform/vault/Dockerfile.offline tests/platform/vault/
 	docker run --network ${networkName} -v $$(pwd):/mnt -td -p 8125:8125 -p 8200:8200 --name ${moduleName} ${moduleImage}  
 
 cleanup:
