+++ tests/modules/loader.py
@@ -1,3 +1,9 @@
+##
+## Python 2.7.X -> 3.0.X
+### Submodules Auto loader
+### Author: Isham J. Araia @ None
+### Date: 10 - 09 - 2018
+
 import os, sys
 lib_path = os.path.abspath(os.path.join('..', '..', 'src'))
 sys.path.append(lib_path)
