--- tests/platform/backend/.gitignore
@@ -1,1 +1,0 @@
-consul
