--- tests/platform/backend/acl.json
@@ -1,6 +1,0 @@
-{
-	"acl_datacenter": "backend-dev-dc",
-    "acl_master_token": "very-secret-token",
-    "acl_default_policy": "deny",
-    "acl_down_policy": "extend-cache"
-}
