+++ tests/platform/consul/.gitignore
@@ -1,0 +1,1 @@
+consul
