+++ tests/platform/consul/acl.json
@@ -1,0 +1,6 @@
+{
+	"acl_datacenter": "backend-dev-dc",
+    "acl_master_token": "very-secret-token",
+    "acl_default_policy": "deny",
+    "acl_down_policy": "extend-cache"
+}
