--- tests/platform/frontend/.gitignore
@@ -1,1 +1,0 @@
-vault
