--- tests/platform/frontend/requirements.pip
@@ -1,1 +1,0 @@
-requests
