+++ tests/platform/vault/.gitignore
@@ -1,0 +1,1 @@
+vault
