+++ tests/platform/vault/Dockerfile.offline
@@ -1,0 +1,8 @@
+FROM centos:7
+
+COPY vault /usr/bin/vault
+COPY vault.hcl /etc/vault.hcl
+RUN yum install nano epel-release -y && \
+	yum install python2-pip nano -y
+
+CMD ["vault", "server", "-config=/etc/vault.hcl"]
