+++ tests/platform/vault/requirements.pip
@@ -1,0 +1,1 @@
+requests
