+++ tests/platform/vault/vault.hcl
@@ -1,0 +1,11 @@
+storage "consul" {
+  address = "consulmod:8500"
+  path    = "vault"
+  token   = "very-secret-token"
+}
+
+listener "tcp" {
+  address     = "127.0.0.1:8200"
+  tls_disable = 1
+  token       = "very-secret-token"
+}
